refactory tests funcionais
