+++ src/funcionaltest/java/br/com/covid/brazil/api/util/RotasEnum.java
