+++ src/funcionaltest/resources/bootstrap-funcional.yml
